Use dropdown for literal identifier lookup
