+++ webapp/scripts/interface.ts
@@ -405,7 +405,10 @@ export function getExportCloseButton(): HTMLButtonElement {
 export function getTreePathOfElement(element: HTMLElement | null): string {
     if (element === null) throw new Error("Cannot get tree path of null");
     const treePath = element.getAttribute("data-tree-path");
-    if (treePath === null) throw new Error("Element does not have a tree path");
+    if (treePath === null) {
+        console.log(element.outerHTML);
+        throw new Error("Element does not have a tree path");
+    }
     return treePath;
 }
