+++ webapp/styles/components/custom_dropdown.scss
@@ -1,0 +1,79 @@
+.dropdown-selector-container {
+  display: inline-flex;
+  position: relative;
+
+  &.focused {
+    box-shadow: -1px 0 2px 1px $foreground, 0 -1px 2px 1px $foreground, 1px 0 2px 1px $foreground, 0 1px 2px 1px $foreground;
+    border-radius: 3px;
+  }
+
+  &.guide-highlight {
+    input, button {
+      background-color: $highlighted-input;
+      color: $highlighted-input-text;
+    }
+  }
+
+  input {
+    margin: 0;
+    border-radius: 3px 0 0 3px;
+    outline: none;
+  }
+
+  button {
+    margin: 0;
+    border: $input-outline solid 1px;
+    border-radius: 0 3px 3px 0;
+  }
+
+  .dropdown {
+    position: absolute;
+    top: 100%;
+    left: 0;
+    background-color: $context-menu-background;
+    color: $foreground;
+    border: $input-outline solid 1px;
+    z-index: 1;
+    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.16), 0 2px 4px rgba(0, 0, 0, 0.23);
+    pointer-events: bounding-box;
+
+    overflow-y: auto;
+    max-height: 20rem;
+
+    &.show {
+      display: block;
+    }
+
+    &:not(.show) {
+      display: none;
+    }
+
+    > ul {
+      list-style: none;
+      padding: 0;
+      margin: 0;
+
+      li {
+        cursor: pointer;
+        padding: 5px;
+        text-align: left;
+        z-index: 100;
+        font-style: normal;
+
+        &.highlight {
+          background-color: $highlighted-tree;
+          box-shadow: 0 2px 4px rgba(0, 0, 0, 0.16), 0 2px 4px rgba(0, 0, 0, 0.23);
+          font-weight: bold;
+        }
+
+        &.hidden {
+          display: none;
+        }
+
+        &:hover {
+          background-color: $hover-background;
+        }
+      }
+    }
+  }
+}
